changed landsat.py name, added pip freeze

+++ metadata.py
@@ -54,11 +54,15 @@ def api_call(code, api_key, **kwargs):
     url =  API_ROOT + code
     kwargs.update(apiKey=api_key)
     params = to_json(**kwargs)
-    response = requests.get(url, params=params).json()
-    if response['error']:
-        raise Exception('EE: {}'.format(response['error']))
+    response = requests.get(url, params=params)
+    if response.status_code != 200:
+        print('Received status code {}'.format(response.status_code))
+    print(response.status_code)
+    data = response.json()
+    if data['error']:
+        raise Exception('EE: {}'.format(data['error']))
     else:
-        return response['data']
+        return data['data']
  
 
 def get_fill(root):
@@ -73,8 +77,9 @@ def return_out_dict(metadata, search_resp, idx):
 
 
 if __name__ == '__main__':
+    api_key = None
     try:
-        if False: #os.path.exists('results.json'):
+        if os.path.exists('results.json'):
             with open('results.json', 'r') as fp:
                 results = json.load(fp)
         else:
@@ -87,30 +92,35 @@ if __name__ == '__main__':
                     sortOrder='DESC',
                     maxResults=1000,
                     additionalCriteria=SEARCH_PARAMS)
+            print('Query returned {} results'.format(len(response['results'])))
             with open('results.json', 'w') as fp:
                 fp.write(json.dumps(response))
 
-        print('Query returned {} results'.format(len(response['results'])))
-        metadata = list(map(lambda r: r['metadataUrl'], response['results']))
-        exc = ThreadPoolExecutor()
-        xml_futs = exc.map(lambda m: requests.get(m).content, metadata)
-        tick = time.time()
-        xmls = [fut for fut in xml_futs]
-        exc.shutdown(wait=False)
-        print('{} s to get all metadata'.format(time.time()-tick))
+        if os.path.exists('mosaic.json'):
+            with open('mosaic.json', 'r') as fp:
+                to_mosaic = json.load(fp)
+        else:
+            metadata = list(map(lambda r: r['metadataUrl'], response['results']))
+            exc = ThreadPoolExecutor()
+            xml_futs = exc.map(lambda m: requests.get(m).content, metadata)
+            tick = time.time()
+            xmls = [fut for fut in xml_futs]
+            exc.shutdown(wait=False)
+            print('{} s to get all metadata'.format(time.time()-tick))
 
-        roots = map(lambda x: ET.fromstring(x), xmls)
-        fills = list(map(get_fill, roots))
-        min_idx = fills.index(min(fills))
-        to_mosaic = []
-        if min_idx > 0 :
-            to_mosaic.append(return_out_dict(metadata, response, min_idx-1))
-        to_mosaic.append(return_out_dict(metadata, response, min_idx))
-        if min_idx+1 < len(metadata):
-            to_mosaic.append(return_out_dict(metadata, response, min_idx+1))
+            roots = map(lambda x: ET.fromstring(x), xmls)
+            fills = list(map(get_fill, roots))
+            min_idx = fills.index(min(fills))
+            to_mosaic = []
+            if min_idx > 0 :
+                to_mosaic.append(return_out_dict(metadata, response, min_idx-1))
+            to_mosaic.append(return_out_dict(metadata, response, min_idx))
+            if min_idx+1 < len(metadata):
+                to_mosaic.append(return_out_dict(metadata, response, min_idx+1))
 
-        with open('mosaic.json', 'w') as fp:
-            fp.write(json.dumps({'mosaic': to_mosaic}))
-            
+            with open('mosaic.json', 'w') as fp:
+                fp.write(json.dumps({'mosaic': to_mosaic}))
+
+                
     except Exception as ex:
         print_exc()
